Backend/requirements.txt
email-validator==2.3.0
fastapi==0.128.0
faster-whisper==1.2.1
huggingface_hub==1.3.4
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.3.31
langchain-core==1.2.7
langchain-google-genai==2.1.12
langchain-groq==1.1.1
langchain-pinecone==0.2.13
pydantic==2.12.5
pydub==0.25.1
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
supabase==2.27.2
uvicorn==0.40.0
